requirements.txt
django>=5.2.1,<5.3
djangorestframework>=3.16.0,<3.17
djangorestframework-simplejwt>=5.5.0,<5.6
psycopg2-binary>=2.9.9,<2.10
django-cors-headers>=4.7.0,<4.8
drf-yasg>=1.21.7,<1.22
python-dotenv>=1.0.0,<1.1
gunicorn>=21.2.0,<21.3
whitenoise>=6.6.0,<6.7
